Add support for Delta 3 Classic
